ON-15724: change method of reading fd in zfstackdump test
Previously, we would redirect fd 1000 into egrep to get the output
of our stackdump, however, this didn't work on newer Ubuntu's and
perhaps others. Instead, we now rely on sysfs to correctly link
to our opened fds.

As supplementary changes, this now uses cat so the output is kept
for the sake of test logs, and grep -E to replace egrep as egrep
is becoming deprecated on some newer OSes.